requirements.txt
crewai
python-dotenv
openai
